Move dyn_struct to helpers module
